styles 🎨 : tighten mobile card and project row padding
