+++ src/components/pages/home/section-work-experience.tsx
@@ -1,6 +1,7 @@
 import type { ReactElement } from 'react';
 
 import { Chip } from '@/components/ui/chip';
+import { EntryCard } from '@/components/ui/entry-card';
 import { Section } from '@/components/ui/section';
 import { SectionHeading } from '@/components/ui/section-heading';
 import { useLocale } from '@/contexts/locale-context';
@@ -17,19 +18,7 @@ export const SectionWorkExperience = () => {
 			/>
 			<div className="flex flex-col gap-[14px]">
 				{data.work.map(({ company, title, description, link, preview, stack, start, end }, i) => (
-					<div
-						key={company}
-						className="entry-card spotlight-border grid grid-cols-[48px_1fr_auto] gap-[18px] p-[22px_22px_22px_20px] border border-border rounded-xl shadow-card"
-						onMouseMove={(e) => {
-							const r = e.currentTarget.getBoundingClientRect();
-							e.currentTarget.style.setProperty('--mouse-x', `${e.clientX - r.left}px`);
-							e.currentTarget.style.setProperty('--mouse-y', `${e.clientY - r.top}px`);
-						}}
-						onMouseLeave={(e) => {
-							e.currentTarget.style.setProperty('--mouse-x', '-999px');
-							e.currentTarget.style.setProperty('--mouse-y', '-999px');
-						}}
-					>
+					<EntryCard key={company}>
 						<div className="w-11 h-11 rounded-[9px] border border-border bg-background grid place-items-center overflow-hidden shrink-0">
 							{preview && (
 								<img
@@ -66,12 +55,12 @@ export const SectionWorkExperience = () => {
 							{start} — {end}
 						</span>
 
-						<p className="m-0 mt-3 font-mono text-[12.5px] leading-[1.7] text-foreground/90 text-pretty [grid-column:2_/-1]">
+						<p className="m-0 mt-3 font-mono text-[12.5px] leading-[1.7] text-foreground/90 text-pretty [grid-column:1_/-1] sm:[grid-column:2_/-1]">
 							{(translations.data.work[i]?.description ?? description).trim()}
 						</p>
 
 						{stack.length > 0 && (
-							<div className="mt-[14px] flex flex-wrap gap-[5px] [grid-column:2_/-1]">
+							<div className="mt-[14px] flex flex-wrap gap-[5px] [grid-column:1_/-1] sm:[grid-column:2_/-1]">
 								{stack.map((st) => {
 									const { label, icon } = st as { label: string; icon: ReactElement };
 									return (
@@ -83,7 +72,7 @@ export const SectionWorkExperience = () => {
 								})}
 							</div>
 						)}
-					</div>
+					</EntryCard>
 				))}
 			</div>
 		</Section>
